+++ src/constants/index.js
@@ -1,1 +1,4 @@
 export const COUNT_ADD = 'COUNT_ADD';
+// consts for Saga asyc actions, probably you do not need this below
+export const SOME_SAGA = 'SOME_SAGA';
+export const SOME_ASYNC_ACTION = 'SOME_ASYNC_ACTION';