+++ src/controller/actions.js
@@ -1,8 +1,13 @@
-import { COUNT_ADD } from '../constants';
+import { COUNT_ADD, SOME_ASYNC_ACTION } from '../constants';
 
 const addCount = count => ({
   type: COUNT_ADD,
   count
 });
 
-export { addCount };
+const someAsyncAction = payload => ({
+  type: SOME_ASYNC_ACTION,
+  payload
+});
+
+export { addCount, someAsyncAction };