+++ src/index.js
@@ -1,24 +1,24 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import RedBox from 'redbox-react';
-import AppContainer from './container/index.js';
+import { syncHistoryWithStore } from 'react-router-redux';
+import createBrowserHistory from 'history/createBrowserHistory';
+
+import store from './controller/store';
+import AppContainer from './container/index';
+
+const history = syncHistoryWithStore(createBrowserHistory(), store);
 
 const ENTRY_POINT = document.querySelector('#react-app-root');
 
 // creating starting endpoint for app.
 const render = () => {
-  ReactDOM.render(
-    <AppContainer />,
-    ENTRY_POINT
-  );
+  ReactDOM.render(<AppContainer store={store} history={history} />, ENTRY_POINT);
 };
 
 // this will help us understand where the problem is located once app will fall.
 const renderError = error => {
-  ReactDOM.render(
-    <RedBox error={error} />,
-    ENTRY_POINT
-  );
+  ReactDOM.render(<RedBox error={error} />, ENTRY_POINT);
 };
 
 // This code is excluded from production bundle
@@ -28,10 +28,7 @@ if (__DEV__) {
   // ========================================================
   const devRender = () => {
     if (module.hot) {
-      module.hot.accept(
-        './container/index.js',
-        () => render()
-      );
+      module.hot.accept('./container/index.js', () => render());
     }
 
     render();